Time: 0 ms (100.00%) | Memory: 42.2 MB (49.20%) - LeetSync

+++ 70-climbing-stairs/climbing-stairs.java
@@ -1,10 +1,10 @@
 class Solution {
     public int climbStairs(int n) {
-        long f = Math.round(
+        int f = (int)(Math.round(
     (Math.pow((1 + Math.sqrt(5)) / 2, n + 1)
    - Math.pow((1 - Math.sqrt(5)) / 2, n + 1))
    / Math.sqrt(5)
-);
-     return (int)f;
+));
+     return f;
     }
 }
